commit ketiga

+++ soal1.c
@@ -30,12 +30,12 @@ int main(){
         // int has_right = 1;
         if (Array[j] == -1){
             if ((Array[j-1] != -1) && (Array[j+1] != -1)){
-                Array[j] = (((double)Array[j-1] + (double)Array[j+1])/2);
+                Array[j] = ((Array[j-1] + Array[j+1])/2);
             }
             else if ((Array[j-1] != -1) && (Array[j+1] == -1)){
                 for (int k = 0; k < N; k++){
                     if (Array[j+k+1] != -1){
-                        Array[j] = (((double)Array[j-1] + (double)Array[j+k+1])/2);
+                        Array[j] = ((Array[j-1] + Array[j+k+1])/2);
                         break;
                     }
                 }
@@ -43,7 +43,7 @@ int main(){
             else if((Array[j-1] == -1) && (Array[j+1] != -1)){
                 for (int k = 0; k < N; k++){
                     if (Array[j-k-1] != -1){
-                        Array[j] = (((double)Array[j-k-1] + (double)Array[j+1])/2);
+                        Array[j] = ((Array[j-k-1] + Array[j+1])/2);
                         break;
                     }
                 }
